+++ rust/main/chains/hyperlane-aleo/src/error.rs
@@ -26,6 +26,29 @@ pub enum HyperlaneAleoError {
     /// Unknown Network
     #[error("Unknown Network with ID: {0}")]
     UnknownNetwork(u16),
+    /// Unknown ISM
+    #[error("Unknown ISM: {0}")]
+    UnknownIsm(String),
+    /// Missing Route
+    #[error("Missing Route: {routing_ism} from origin {origin}")]
+    RoutingIsmMissingRoute {
+        /// The route key
+        routing_ism: String,
+        /// Origin domain
+        origin: u32,
+    },
+    /// Mailbox uninitialized
+    #[error("Mailbox uninitialized")]
+    MailboxUninitialized,
+    /// App uninitialized
+    #[error("App uninitialized")]
+    AppUninitialized,
+    /// Unknown Merkle Tree Hook
+    #[error("Unknown Merkle Tree Hook: {0}")]
+    UnknownMerkleTreeHook(String),
+    /// TryFromSliceError
+    #[error("{0}")]
+    TryFromSliceError(#[from] std::array::TryFromSliceError),
     /// Other errors
     #[error("{0}")]
     Other(String),