+++ rust/main/chains/hyperlane-aleo/src/lib.rs
@@ -10,15 +10,25 @@ mod config;
 mod error;
 mod indexer;
 mod ism;
+mod mailbox;
 mod provider;
 mod signer;
 mod types;
 mod utils;
+mod validator_announce;
 
 pub(crate) use types::*;
 
-pub use {
-    config::*, error::*, indexer::AleoDeliveryIndexer, indexer::AleoDispatchIndexer,
-    indexer::AleoInterchainGasIndexer, indexer::AleoMerkleTreeHook, ism::AleoIsm,
-    provider::AleoProvider, signer::AleoSigner,
+pub use config::*;
+pub use error::*;
+pub use indexer::{
+    AleoDeliveryIndexer, AleoDispatchIndexer, AleoInterchainGasIndexer, AleoMerkleTreeHook,
 };
+pub use ism::AleoIsm;
+pub use mailbox::AleoMailbox;
+pub use provider::{AleoProvider, AleoProviderForLander};
+pub use signer::AleoSigner;
+pub use types::{AleoGetMappingValue, AleoTxData, FeeEstimate};
+pub use validator_announce::AleoValidatorAnnounce;
+
+pub use snarkvm::prelude::Plaintext;