+++ rust/main/lander/src/adapter/chains.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "aleo")]
+pub use aleo::AleoTxPrecursor;
 pub use ethereum::EthereumTxPrecursor;
 pub use factory::AdapterFactory;
 pub use radix::RadixTxPrecursor;
@@ -6,7 +8,12 @@ pub use sealevel::SealevelTxPrecursor;
 mod factory;
 
 // chains modules below
+#[cfg(feature = "aleo")]
+mod aleo;
 mod cosmos;
 pub mod ethereum;
 pub mod radix;
 pub mod sealevel;
+
+#[cfg(all(test, feature = "aleo"))]
+pub use aleo::AleoAdapter;