+++ rust/main/lander/src/adapter/chains/aleo/adapter.rs
@@ -1,0 +1,5 @@
+pub use core::AleoAdapter;
+
+mod build;
+mod core;
+mod submit;