+++ rust/main/lander/src/adapter/chains/radix.rs
@@ -2,5 +2,6 @@ pub mod adapter;
 pub mod conf;
 
 mod precursor;
+mod transaction;
 
 pub use precursor::*;