+++ rust/main/lander/src/tests.rs
@@ -1,5 +1,11 @@
+#[cfg(all(test, feature = "aleo"))]
+pub use aleo::MockAleoProvider;
+
+#[cfg(feature = "aleo")]
+mod aleo;
 pub mod evm;
 pub mod radix;
 pub mod svm;
+
 #[cfg(test)]
 pub mod test_utils;