+++ rust/main/lander/src/tests/aleo.rs
@@ -1,0 +1,8 @@
+#[cfg(test)]
+pub use test_utils::MockAleoProvider;
+
+#[cfg(test)]
+mod test_utils;
+
+#[cfg(test)]
+mod tests_building_stage;