+++ typescript/provider-sdk/src/chain.ts
@@ -9,11 +9,14 @@ type ChainNameOrId = string | number;
  */
 export interface ChainMetadataForAltVM {
   name: string;
+  bech32Prefix?: string;
   domainId: Domain;
   chainId: ChainId;
   nativeToken?: {
     decimals?: number;
     denom?: string;
+    symbol: string;
+    name: string;
   };
   blocks?: {
     confirmations?: number;
@@ -22,6 +25,22 @@ export interface ChainMetadataForAltVM {
   rpcUrls?: {
     http: string;
   }[];
+  gatewayUrls?: {
+    http: string;
+  }[];
+  // Used in radix tests after deploying
+  // the package address to the local chain
+  packageAddress?: string;
+}
+
+/**
+ * Test chain metadata type with additional fields for testing
+ */
+export interface TestChainMetadata extends ChainMetadataForAltVM {
+  rpcPort: number;
+  rpcUrl: string;
+  restPort: number;
+  gnosisSafeTransactionServiceUrl?: string;
 }
 
 /**