+++ typescript/radix-sdk/src/utils/constants.ts
@@ -1,0 +1,13 @@
+/**
+ * Radix public key to be used for read/simulation in contexts where a
+ * signer is not available.
+ *
+ * Generated using:
+ * ```typescript
+ *
+ * const pk = new PrivateKey.Ed25519(new Uint8Array(randomBytes(32)));
+ * const pubKey = pk.publicKeyHex()
+ * ```
+ */
+export const READ_ACCOUNT_HEX_PUBLIC_KEY =
+  '4e87d816766ad391ed7e91e09395b5ff1d899be0b94df63f6f2301acf7966dbc';